+++ case-elider/index.html
@@ -3,7 +3,7 @@
   <head>
     <title>Case elision demo</title>
     <meta charset="utf-8" />
-    <script src="api.js" defer></script>
+    <script src="api.js" type="module"></script>
     <link rel="stylesheet" href="styles.css" />
   </head>
   <body>
